+++ CHANGELOG.md
@@ -36,6 +36,7 @@ release.
 ## [Unreleased]
 
 ### Changed
+- Added `limitCk` and `limitSpk` query params to limit the number of CK and SPK files returned [#82](https://github.com/DOI-USGS/SpiceQL/pull/82)
 - Turned `limitQuality` into `limitCk` and `limitSpk` query params [#80](https://github.com/DOI-USGS/SpiceQL/pull/80)
 
 ## 1.2.0